requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
requests>=2.31.0
groq>=0.9.0
python-dotenv>=1.0.0